Feature/usecase/receiver (#10)
* impl mock consensus sender

* add TxValidate StatelessValidator

* client gate receiver test

* WIP lock clearly memory

* impl logic

* fix logk

* check proposal tx queue already exist

* impl receiver pool

* WIP : impl DOS test;

* impl Propagate receiver and test

* fix lock

* impl propose receiver

* impl propose receiver test

* impl vote receiver

* impl vote test

* add receiveChannel

* in circleci small case

* fix checkAndLock

* use capacity on lock

* add read mutex lock

* no use cap;

+++ model/factory.go
@@ -9,7 +9,7 @@ var (
 
 type ModelFactory interface {
 	NewBlock(height int64, preBlockHash []byte, createdTime int64, txs []Transaction) (Block, error)
-	NewProposal(block Block, round int64) (Proposal, error)
+	NewProposal(block Block, round int32) (Proposal, error)
 	NewVoteMessage(hash []byte) VoteMessage
 	NewSignature(pubkey []byte, signature []byte) Signature
 	NewPeer(address string, pubkey []byte) Peer
